requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
frozenlist==1.3.3
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
oauthlib==3.2.2
openai==0.26.2
python-dotenv==0.21.1
requests==2.28.2
requests-async==0.6.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
tqdm==4.64.1
tweepy==4.12.1
urllib3==1.26.14
yarl==1.8.2
